move combined_datasets.get_timeseries_for_fips columns and min_range_with_some_value logic to TimeseriesDataset.get_columns_and_date_subset so it can be run on data not read with combined_datasets
