+++ test/combined_dataset_test.py
@@ -134,27 +134,6 @@ def test_expected_field_in_sources(data_source_cls):
         assert len(good_state) >= 48
 
 
-@pytest.mark.parametrize("include_na_at_end", [False, True])
-def test_remove_padded_nans(include_na_at_end):
-    rows = [
-        {"date": "2020-02-01", "cases": pd.NA},
-        {"date": "2020-02-02", "cases": pd.NA},
-        {"date": "2020-02-03", "cases": 1},
-        {"date": "2020-02-04", "cases": pd.NA},
-        {"date": "2020-02-05", "cases": 2},
-        {"date": "2020-02-06", "cases": 3},
-    ]
-    if include_na_at_end:
-        rows += [{"date": "2020-02-07", "cases": pd.NA}]
-
-    df = pd.DataFrame(rows)
-
-    results = combined_datasets._remove_padded_nans(df, ["cases"])
-    expected_series = pd.Series([1, pd.NA, 2, 3], name="cases")
-
-    pd.testing.assert_series_equal(results.cases, expected_series)
-
-
 def test_build_timeseries():
     data_a = read_csv_and_index_fips_date(
         "county,state,fips,country,aggregate_level,date,cases\n"
