+++ test/libs/datasets/timeseries_test.py
@@ -1,0 +1,25 @@
+import pytest
+import pandas as pd
+
+from libs.datasets import timeseries
+
+
+@pytest.mark.parametrize("include_na_at_end", [False, True])
+def test_remove_padded_nans(include_na_at_end):
+    rows = [
+        {"date": "2020-02-01", "cases": pd.NA},
+        {"date": "2020-02-02", "cases": pd.NA},
+        {"date": "2020-02-03", "cases": 1},
+        {"date": "2020-02-04", "cases": pd.NA},
+        {"date": "2020-02-05", "cases": 2},
+        {"date": "2020-02-06", "cases": 3},
+    ]
+    if include_na_at_end:
+        rows += [{"date": "2020-02-07", "cases": pd.NA}]
+
+    df = pd.DataFrame(rows)
+
+    results = timeseries._remove_padded_nans(df, ["cases"])
+    expected_series = pd.Series([1, pd.NA, 2, 3], name="cases")
+
+    pd.testing.assert_series_equal(results.cases, expected_series)
